refactor page header to use sticky over fixed
The sticky positioning approach is more maintainable and performs better than the previous implementation using fixed
positioning with dynamic height calculations. It also performs better with no js.

- shared `siteHeaderHeight` CSS variable to standardize header heights
- remove the complex buffer div that was previously needed with fixed positioning
- eliminate the janky dependency on `useElementSize` hook for header height calculations

This approach is also better than "cloning the element and setting the second to invisible". Technically, it's only
better because of semantics (no repeating content, esp h1). It's not as reusable/ foolproof if I didn't know the header
height ahead of time.